third_party/blink/web_tests/wpt_internal/webmcp/initial-about-blank-shared-tool.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="author" href="mailto:dom@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
// TODO(https://crbug.com/506393880): Once the official Promise-returning
// `getTools()` API exists, we can move this test from `wpt_internal` to
// `external/wpt/webmcp`.
promise_test(async t => {
  const iframe = document.createElement('iframe');
  iframe.src = '/';
  document.body.append(iframe);
  // Register a tool with the initial about:blank Document [1].
  //
  // [1]: https://html.spec.whatwg.org/#is-initial-about:blank.
  console.log(iframe.contentWindow.navigator.modelContext);
  iframe.contentWindow.navigator.modelContext.registerTool({
    name: "default-tool",
    description: "default desc",
    execute: async () => "hello"
  });
  assert_equals(iframe.contentDocument.URL, "about:blank");
  assert_equals(
      iframe.contentWindow.navigator.modelContextTesting.listTools()[0].name,
      "default-tool");

  // Wait for the iframe navigation to complete.
  await new Promise(resolve => iframe.onload = resolve);
  assert_not_equals(iframe.contentDocument.URL, "about:blank");
  // Note that the tool still incorrectly persists, even though we've navigated
  // away from the initial about:blank Document.
  assert_array_equals(
      iframe.contentWindow.navigator.modelContextTesting.listTools(), []);
}, "tools registered with the initial about:blank Document do not survive navigation");
</script>
</body>
